ImmuneAppTransPHLAService/requirements.txt
absl-py==2.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
certifi==2022.12.7
cffi==1.15.1
cycler==0.11.0
fonttools==4.38.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.62.3
h5py==2.10.0
importlib-metadata==6.7.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
logomaker==0.8
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
minio==7.2.7
numpy==1.20.0
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.3
pathlib==1.0.1
Pillow==9.5.0
protobuf==3.20.0
pycparser==2.21
pycryptodome==3.22.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
scipy==1.7.2
seaborn==0.11.2
six==1.17.0
tensorboard==1.15.0
tensorflow==1.15
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
scikit-learn==0.24.2
torch==1.7.1
tqdm==4.54.1
uvicorn==0.22.0
fastapi==0.103.2
requests==2.31.0
python-dotenv==0.21.1